pyproject.toml
[project]
authors = [
    {name = "dbt Labs", email = "info@dbtlabs.com"},
]
license = {text = "Apache License 2.0"}
requires-python = ">=3.9"
dependencies = [
    "click<9.0.0,>=8.1.3",
    "requests<3.0.0,>=2.32.0",
    "loguru<1.0.0,>=0.6.0",
    "deepdiff>=8.0.0,<9.0.0",
    "pydantic<3.0.0,>=2.0.0",
    "croniter<2.0.0,>=1.3.8",
    "ruamel-yaml<1.0.0,>=0.17.21",
    "rich<13.0.0,>=12.6.0",
    "PyYAML<7.0.0,>=6.0.1",
    "python-dateutil<3.0,>=2.9",
    "beartype<1.0.0,>=0.18.5",
    "jinja2>=3.1.5,<4.0.0",
    "importlib-metadata<7,>=6.0",
]
name = "custom-dbt-jobs-as-code"
dynamic = ["version"]
description = "Custom dbt jobs as code CLI - for specific requirements"
readme = "README.md"
keywords = [
    "dbt",
    "dbt Cloud",
]

[project.urls]
repository = "https://github.com/mani-dbt/custom-dbt-jobs-as-code"

[project.scripts]
custom-dbt-jobs-as-code = "dbt_jobs_as_code.main:cli"

[dependency-groups]
dev = [
    "coverage<8.0.0,>=7.6.3",
    "jsonschema<5.0.0,>=4.17.3",
    "pytest-mock>=3.14.0",
    "pytest<8.0.0,>=7.2.0",
    "pytest-beartype<1.0.0,>=0.0.2",
    "pytest-cov<6.0.0,>=5.0.0",
    "pre-commit",
]
mkdocs = [
    "mkdocs-click>=0.8.1",
    "mkdocs-material>=9.6.3",
    "mike>=2.1.3",
    "mkdocs-git-revision-date-localized-plugin>=1.3.0",
]

[tool.pdm.build]
includes = ["src/dbt_jobs_as_code"]

[tool.pdm.version]
source = "scm"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
beartype_packages = 'src'
markers = [
    "not_in_parallel: marks tests that can't run in parallel",
]

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"] # line-length, need to fix those by hands first
typing-modules = ["beartype.typing"]